refactor: relocate unit tests
