+++ tests/Sinch.Tests/Conversation/WebhooksTests.cs
@@ -17,71 +17,6 @@ namespace Sinch.Tests.Conversation
 {
     public class WebhooksTests : ConversationTestBase
     {
-        private const string TimestampHeader = "x-sinch-webhook-signature-timestamp";
-        private const string NonceHeader = "x-sinch-webhook-signature-nonce";
-        private const string AlgorithmHeader = "x-sinch-webhook-signature-algorithm";
-        private const string SignatureHeader = "x-sinch-webhook-signature";
-
-        [Fact]
-        public void ValidateRequest()
-        {
-            var json = Helpers.LoadResources("Conversation/Hooks/WebhooksAuthValidationSingleLine.json");
-
-            var isValid = Conversation.Webhooks.ValidateAuthenticationHeader(new Dictionary<string, IEnumerable<string>>()
-            {
-                { NonceHeader, ["01FJA8B4A7BM43YGWSG9GBV067"] },
-                { TimestampHeader, ["1634579353"] },
-                { AlgorithmHeader, ["HmacSHA256"] },
-                { SignatureHeader, ["6bpJoRmFoXVjfJIVglMoJzYXxnoxRujzR4k2GOXewOE="] },
-            }, json, "foo_secret1234");
-            isValid.Should().BeTrue();
-        }
-
-        [Fact]
-        public void ValidateRequest_WithSingleValueHeaders()
-        {
-            var json = Helpers.LoadResources("Conversation/Hooks/WebhooksAuthValidationSingleLine.json");
-
-            var isValid = Conversation.Webhooks.ValidateAuthenticationHeader(new Dictionary<string, string>()
-            {
-                { NonceHeader, "01FJA8B4A7BM43YGWSG9GBV067" },
-                { TimestampHeader, "1634579353" },
-                { AlgorithmHeader, "HmacSHA256" },
-                { SignatureHeader, "6bpJoRmFoXVjfJIVglMoJzYXxnoxRujzR4k2GOXewOE=" },
-            }, json, "foo_secret1234");
-            isValid.Should().BeTrue();
-        }
-
-        [Fact]
-        public void ValidateMultiLinePayloadRequest()
-        {
-            var json = Helpers.LoadResources("Conversation/Hooks/WebhooksAuthValidation.json");
-
-            var isValid = Conversation.Webhooks.ValidateAuthenticationHeader(new Dictionary<string, IEnumerable<string>>()
-            {
-                { NonceHeader, ["01FJA8B4A7BM43YGWSG9GBV067"] },
-                { TimestampHeader, ["1634579353"] },
-                { AlgorithmHeader, ["HmacSHA256"] },
-                { SignatureHeader, ["wKmZBGo4Cf+y9cZoPHhiVw6ziKeubLGqN4OdG8jlaPo="] },
-            }, json, "foo_secret1234");
-            isValid.Should().BeTrue();
-        }
-
-        [Fact]
-        public void InValidateMultiLinePayloadRequest()
-        {
-            var json = Helpers.LoadResources("Conversation/Hooks/WebhooksAuthValidation.json");
-
-            var isValid = Conversation.Webhooks.ValidateAuthenticationHeader(new Dictionary<string, IEnumerable<string>>()
-            {
-                { NonceHeader, ["01FJA8B4A7BM43YGWSG9GBV067"] },
-                { TimestampHeader, ["1634579353"] },
-                { AlgorithmHeader, ["HmacSHA256"] },
-                { SignatureHeader, ["wKmZBGo4Cf+y9cZoPHhiVw6ziKeubLGqN4OdG8jlaPo="] },
-            }, json, "wrong_secret");
-            isValid.Should().BeFalse();
-        }
-
         [Fact]
         public void DeserializeCapabilityEvent()
         {
